+++ docs/test_only.py
@@ -1,0 +1,16 @@
+#comment lines to run tests on individual rst files
+
+import doctest
+doctest.testfile("legsTutorial.rst")
+#doctest.testfile("acknowledgement.rst")
+#doctest.testfile("contribute.rst")
+#doctest.testfile("geoDoc.rst")
+#doctest.testfile("index.rst")
+#doctest.testfile("install.rst")
+#doctest.testfile("legsDoc.rst")
+#doctest.testfile("legsResources.rst")
+#doctest.testfile("legsTutorial.rst")
+#doctest.testfile("radarDoc.rst")
+#doctest.testfile("radarTutorial.rst")
+#doctest.testfile("radarTutorialIndex.rst")
+#